# sencondhandproject/api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "sencondHandProject Api Documentation",
    "description": "sencondHandProject Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "sencondHandProject Logo",
      "altText": "sencondHandProject Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "sencondHandProject API Support",
      "url": "https://sencondhandproject.com/info/support",
      "email": "support@sencondhandproject.com"
    },
    "termsOfService": "https://sencondhandproject.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/android": {
      "get": {
        "summary": "List Android",
        "description": "Android",
        "operationId": "ListAndroid",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Android"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAndroid"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/iphone": {
      "get": {
        "summary": "List Iphone",
        "description": "Iphone",
        "operationId": "ListIphone",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Iphone"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListIphone"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/account": {
      "get": {
        "summary": "User Account Details",
        "description": "My Account",
        "operationId": "Account",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Account"
        ],
        
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthorized"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ListAndroid": {
        "type": "object",
        "properties": {
          "brand": {
            "description": "Brand",
            "type": "string"
          },
          "model": {
            "description": "Model",
            "type": "string"
          },
          "price": {
            "description": "Price",
            "type": "string"
          },
          "owner": {
            "description": "Owner",
            "type": "string"
          },
          "transactiontime": {
            "description": "Transactiontime",
            "type": "string"
          },
          "issold": {
            "description": "Issold",
            "type": "string"
          },
          "imagepath": {
            "description": "Imagepath",
            "type": "string"
          }
        }
      },
...
      "ListIphone": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "brand": {
            "description": "Brand",
            "type": "string"
          },
          "model": {
            "description": "Model",
            "type": "string"
          },
          "price": {
            "description": "Price",
            "type": "string"
          },
          "owner": {
            "description": "Owner",
            "type": "string"
          },
          "transactiontime": {
            "description": "Transactiontime",
            "type": "string"
          },
          "issold": {
            "description": "Issold",
            "type": "string"
          },
          "imagepath": {
            "description": "Imagepath",
            "type": "string"
          }
        }
      },
...
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "email": {
            "description": "Email",
            "type": "string"
          },
          "address": {
            "description": "Address",
            "type": "string"
          },
          "email_verified_at": {
            "description": "Email Verified At",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "token": {
            "description": "Token",
            "type": "string"
          },
          "user_role_id": {
            "description": "User Role Id",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
...
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
      "UnAuthorized": {
        "type": "object",
        "properties": {
          "code": {
            "description": "403",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthorized",
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}